requirements.txt
# Core dependencies
torch>=1.7.0
torchvision>=0.8.1
ultralytics>=8.0.0
opencv-python>=4.5.1
numpy>=1.19.0
pandas>=1.2.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Jupyter and visualization
jupyter>=1.0.0
ipykernel>=5.3.4
ipywidgets>=7.6.3

# Utilities
tqdm>=4.50.0
pillow>=8.0.0
scipy>=1.6.0
pyyaml>=5.3.1

# Testing
pytest>=6.0.0
pytest-cov>=2.10.0
